Run r10k bootstrap only when Puppetfile was changed

+++ puppet-bootstrap.sh
@@ -1,8 +1,13 @@
 #!/bin/sh
 
-apt-get update -qq
-[ -x /usr/bin/git ] || apt-get install -y -q git
-[ -x /usr/bin/gem ] || apt-get install -y -q rubygems
-[ -x /usr/local/bin/r10k ] || gem install --no-rdoc --no-ri r10k
-cd /vagrant/vagrant && r10k -v info puppetfile install
+cd /vagrant/vagrant
 
+if [ ! -e modules/.r10k_stamp ] || [ modules/.r10k_stamp -ot Puppetfile ]; then
+  apt-get update -qq
+  [ -x /usr/bin/git ] || apt-get install -y -q git
+  [ -x /usr/bin/gem ] || apt-get install -y -q rubygems
+  [ -x /usr/local/bin/r10k ] || gem install --no-rdoc --no-ri r10k
+
+  r10k -v info puppetfile install
+  touch modules/.r10k_stamp
+fi
